Changed residiential and commercial to deal with userId, added agentId to claims

+++ DEPI-PROJECT.BLL/Services/Implements/CommercialPropertyService.cs
@@ -31,13 +31,14 @@ namespace DEPI_PROJECT.BLL.Services.Implements
             _likePropertyRepo = likePropertyRepo;
         }
 
-        public async Task<ResponseDto<PagedResultDto<CommercialPropertyReadDto>>> GetAllPropertiesAsync(Guid CurrentUserId, CommercialPropertyQueryDto queryDto)
+        public async Task<ResponseDto<PagedResultDto<CommercialPropertyReadDto>>> GetAllPropertiesAsync(Guid UserId, CommercialPropertyQueryDto queryDto)
         {
-            var query = _repo.GetAllProperties();
-
+            IQueryable<CommercialProperty> query = _repo.GetAllProperties();
+            
             var result = await query.IF(queryDto.BusinessType != null, a => a.BusinessType.Contains(queryDto.BusinessType))
                                     .IF(queryDto.FloorNumber != null, a => a.FloorNumber == queryDto.FloorNumber)
                                     .IF(queryDto.HasStorage != null, a => a.HasStorage == queryDto.HasStorage)
+                                    .IF(queryDto.UserId != null, a => a.Agent.UserId == queryDto.UserId)
                                     .Paginate(new PagedQueryDto { PageNumber = queryDto.PageNumber, PageSize = queryDto.PageSize })
                                         .ToListAsync();
 
@@ -57,7 +58,7 @@ namespace DEPI_PROJECT.BLL.Services.Implements
                                     .ToDictionaryAsync(n => n.PropertyId, n => n.Count);
 
             var IsLikedHash = await _likePropertyRepo.GetAllLikesByPropertyIds(PropertiesIds)
-                                    .Where(lc => lc.UserID == CurrentUserId)
+                                    .Where(lc => lc.UserID == UserId)
                                     .Select(n => n.PropertyId)
                                     .ToHashSetAsync();
 
@@ -79,7 +80,7 @@ namespace DEPI_PROJECT.BLL.Services.Implements
                 {
                     property.IsLiked = false;
                 }
-            } 
+            }
             #endregion
 
             var pagedResult = new PagedResultDto<CommercialPropertyReadDto>(mappedData, queryDto.PageNumber, query.Count(), queryDto.PageSize);
@@ -92,7 +93,7 @@ namespace DEPI_PROJECT.BLL.Services.Implements
             };
         }
 
-        public async Task<ResponseDto<CommercialPropertyReadDto>> GetPropertyByIdAsync(Guid CurrentUserId, Guid id)
+        public async Task<ResponseDto<CommercialPropertyReadDto>> GetPropertyByIdAsync(Guid UserId, Guid id)
         {
             var property = await _repo.GetPropertyByIdAsync(id);
             if (property == null)
@@ -101,12 +102,12 @@ namespace DEPI_PROJECT.BLL.Services.Implements
             }
             var mapped = _mapper.Map<CommercialPropertyReadDto>(property);
 
-            // #region Add Islike , count for each comment
-            // //count likes --> call likeCommentRepo
-            // mapped.LikesCount = await _likePropertyRepo.CountLikesByPropertyId(mapped.PropertyId);
-            // //check is liked by Current user
-            // mapped.IsLiked = await _likePropertyRepo.GetLikePropertyByUserAndPropertyId(CurrentUserId, mapped.PropertyId) != null; 
-            // #endregion
+            #region Add Islike , count for each comment
+            //count likes --> call likeCommentRepo
+            mapped.LikesCount = await _likePropertyRepo.CountLikesByPropertyId(mapped.PropertyId);
+            //check is liked by Current user
+            mapped.IsLiked = await _likePropertyRepo.GetLikePropertyByUserAndPropertyId(UserId, mapped.PropertyId) != null; 
+            #endregion
 
             return new ResponseDto<CommercialPropertyReadDto>
             {
@@ -116,12 +117,12 @@ namespace DEPI_PROJECT.BLL.Services.Implements
             };
         }
 
-        public async Task<ResponseDto<bool>> UpdateCommercialPropertyAsync(Guid id, CommercialPropertyUpdateDto propertyDto)
+        public async Task<ResponseDto<bool>> UpdateCommercialPropertyAsync(Guid UserId, Guid id, CommercialPropertyUpdateDto propertyDto)
         {
             var existing = await _repo.GetPropertyByIdAsync(id);
-            if (existing == null)
+            if (existing == null || existing.Agent.UserId != UserId)
             {
-                throw new NotFoundException($"No property found with ID {id}");
+                throw new NotFoundException($"No property found with ID {id} for UserId {UserId}");
             }
             _mapper.Map(propertyDto, existing);
             if (propertyDto.Amenity != null)
@@ -146,9 +147,15 @@ namespace DEPI_PROJECT.BLL.Services.Implements
                 Data = true
             };
         }
-        public async Task<ResponseDto<CommercialPropertyReadDto>> AddPropertyAsync(CommercialPropertyAddDto propertyDto)
+        public async Task<ResponseDto<CommercialPropertyReadDto>> AddPropertyAsync(Guid UserId, Guid AgentId, CommercialPropertyAddDto propertyDto)
         {
+            if(propertyDto.UserId != UserId)
+            {
+                throw new UnauthorizedAccessException($"Current user unauthorized to do such action, mismatch Ids: Current ID {UserId}, givenId {propertyDto.UserId}");
+            }
             var property = _mapper.Map<CommercialProperty>(propertyDto);
+            property.AgentId = AgentId;
+
             await _repo.AddCommercialPropertyAsync(property);
             if (propertyDto.Amenity != null)
             {
@@ -156,20 +163,28 @@ namespace DEPI_PROJECT.BLL.Services.Implements
                 amenity.PropertyId = property.PropertyId;
                 await _repo.AddAmenityAsync(amenity);
             }
+
+            var PropertyResponseDto = _mapper.Map<CommercialPropertyReadDto>(property);
+            PropertyResponseDto.UserId = propertyDto.UserId;
+
             return new ResponseDto<CommercialPropertyReadDto>
             {
                 IsSuccess = true,
                 Message = "Commercial property added successfully.",
-                Data = _mapper.Map<CommercialPropertyReadDto>(property)
+                Data = PropertyResponseDto
             };
         }
 
-        public async Task<ResponseDto<bool>> DeleteCommercialPropertyAsync(Guid id)
+        public async Task<ResponseDto<bool>> DeleteCommercialPropertyAsync(Guid UserId, Guid id)
         {
             var existing = await _repo.GetPropertyByIdAsync(id);
             if (existing == null)
             {
-                throw new NotFoundException($"No property found with ID {id}");
+                throw new NotFoundException($"No property found with ID {id} for userId {UserId}");
+            }
+
+            if(existing.Agent.UserId != UserId){
+                throw new UnauthorizedAccessException($"Mismatch user Ids: Current {UserId}, given {existing.Agent.UserId}");
             }
 
             await _repo.DeleteCommercialPropertyAsync(id);
